Move Routing.kt to the other settings since: it extends Application class, and it messes up with the concrete routes.
