+++ backend/src/test/kotlin/com/example/helpers/TestExtensions.kt
@@ -3,7 +3,7 @@ package com.example.helpers
 import com.example.config.initDatabase
 import com.example.repository.TextRepositoryImpl
 import com.example.repository.UserRepositoryImpl
-import com.example.routes.configureRouting
+import com.example.config.configureRouting
 import com.example.service.TextServiceImpl
 import com.example.service.UserServiceImpl
 import io.ktor.server.application.Application
